+++ src/routes/ViewJobLog.js
@@ -301,7 +301,7 @@ export class ViewJobLog extends Component {
       relatedHoldingsInfo,
       relatedItemInfo,
       relatedAuthorityInfo,
-      relatedOrderInfo,
+      relatedPoLineInfo,
       relatedInvoiceInfo,
       relatedInvoiceLineInfo,
     } = record;
@@ -312,7 +312,7 @@ export class ViewJobLog extends Component {
       [OPTIONS.HOLDINGS]: relatedHoldingsInfo.error || '',
       [OPTIONS.ITEM]: relatedItemInfo.error || '',
       [OPTIONS.AUTHORITY]: relatedAuthorityInfo.error || '',
-      [OPTIONS.ORDER]: relatedOrderInfo.error || '',
+      [OPTIONS.ORDER]: { poLineInfo: relatedPoLineInfo.error || '' },
       [OPTIONS.INVOICE]: {
         invoiceInfo: relatedInvoiceInfo.error || '',
         invoiceLineInfo: relatedInvoiceLineInfo?.error || '',