+++ src/settings/MappingProfiles/MappingProfilesForm.js
@@ -2,7 +2,6 @@ import React, {
   useMemo,
   useState,
   useEffect,
-  useLayoutEffect,
 } from 'react';
 import PropTypes from 'prop-types';
 import {
@@ -14,10 +13,7 @@ import {
   Field,
   change,
 } from 'redux-form';
-import {
-  isEmpty,
-  omit,
-} from 'lodash';
+import { omit } from 'lodash';
 
 import stripesForm from '@folio/stripes/form';
 import {
@@ -125,42 +121,51 @@ export const MappingProfilesFormComponent = ({
   const isSubmitDisabled = pristine || submitting;
 
   const [folioRecordType, setFolioRecordType] = useState(existingRecordType || null);
+  const [initials, setInitials] = useState(profile || {});
   const [fieldMappingsForMARCSelectedOption, setFieldMappingsForMARCSelectedOption] = useState('');
   const [fieldMappingsForMARC, setFieldMappingsForMARC] = useState(marcMappingOption || '');
   const [addedRelations, setAddedRelations] = useState([]);
   const [deletedRelations, setDeletedRelations] = useState([]);
-  const [prevExistingRecordType, setPrevExistingRecordType] = useState(existingRecordType);
-  const [initials, setInitials] = useState({
-    ...profile,
-    mappingDetails: isEmpty(mappingDetails) ? getInitialDetails(prevExistingRecordType, true) : mappingDetails,
-  });
   const [isConfirmEditModalOpen, setConfirmModalOpen] = useState(false);
 
-  useLayoutEffect(() => {
-    const isEqual = folioRecordType === prevExistingRecordType;
-    const needsUpdate = !id || (id && (!isEqual || isEmpty(mappingDetails)));
+  const setMappingDetailsToForm = newMappingDetails => {
+    dispatch(change(formName, 'profile.mappingDetails', newMappingDetails));
+  };
 
-    if (!needsUpdate) {
-      return;
-    }
+  // set initial mappingDetails
+  useEffect(() => {
+    // existingRecordType is the initial folio record type value
+    setFolioRecordType(existingRecordType);
 
+    // on Edit/Duplicate mappingDetails is an object of fields
+    // on Create mappingDetails = {}
     const initialMappingDetails = isDuplicateMode ? getMappingDetailsForDuplicated(mappingDetails) : mappingDetails;
-    const newInitDetails = folioRecordType === existingRecordType && !isEmpty(mappingDetails)
-      ? initialMappingDetails
-      : getInitialDetails(folioRecordType, true);
 
+    // update initials
     const newInitials = {
       ...initials,
-      mappingDetails: newInitDetails,
+      mappingDetails: initialMappingDetails,
     };
 
     setInitials(newInitials);
-    dispatch(change(formName, 'profile.mappingDetails', newInitDetails));
 
-    if (!isEqual) {
-      setPrevExistingRecordType(folioRecordType);
+    // initialize the form with mappingDetails
+    setMappingDetailsToForm(initialMappingDetails);
+  }, [existingRecordType]); // eslint-disable-line react-hooks/exhaustive-deps
+
+  // set updated mappingDetails on the folio record type change
+  useEffect(() => {
+    const isInitial = folioRecordType === existingRecordType;
+
+    if (isInitial) {
+      setMappingDetailsToForm(initials.mappingDetails);
+    } else {
+      const updatedMappingDetails = getInitialDetails(folioRecordType, true);
+
+      setMappingDetailsToForm(updatedMappingDetails);
     }
   }, [folioRecordType]); // eslint-disable-line react-hooks/exhaustive-deps
+
   useEffect(() => {
     dispatch(change(formName, 'addedRelations', addedRelations));
   }, [addedRelations, dispatch]);